Fix some windows bugs

+++ Cheukoin/Card.cpp
@@ -98,7 +98,7 @@ bool Card::isGreaterThan(Card other, Suit askedSuit) const
         // 2 cards are from same suit
         map<Value, int> order = _suit == asset ? Rules::CardValuesAsset : Rules::CardValues;
         if (order[_value] == order[other.getValue()]) {
-            return _value < other.getValue();
+            return (int)_value < (int)other.getValue();
         }
         return order[_value] > order[other.getValue()];
     }
